requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
audioop-lts==0.2.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
boto3==1.40.9
botocore==1.40.9
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
doclayout_yolo==0.0.4
emmet-core==0.84.9
fast-langdetect==0.2.5
fastapi==0.116.1
fasttext-predict==0.9.2.4
ffmpy==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
google==3.0.0
google-auth==2.40.3
google-genai==1.29.0
gradio==5.42.0
gradio_client==1.11.1
gradio_pdf==0.0.22
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
invoke==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json_repair==0.49.0
jsonlines==4.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
latexcodec==3.0.1
loguru==0.7.3
maggma==0.72.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mineru==2.1.10
modelscope==1.28.2
mongomock==4.3.0
monty==2025.3.3
mp-api==0.45.8
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
narwhals==2.1.2
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.99.9
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.2
outcome==1.3.0.post0
packaging==25.0
palettable==3.3.3
pandas==2.3.1
paramiko==4.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pdftext==0.6.3
pillow==11.3.0
plotly==6.3.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybtex==0.25.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydash==8.0.5
pydub==0.25.1
Pygments==2.19.2
pymatgen==2025.6.14
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
rapid-table==1.0.5
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
rich==14.1.0
robust-downloader==0.0.2
rpds-py==0.27.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.8
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.18.1
semantic-version==2.10.0
sentinels==1.1.1
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spglib==2.6.0
sshtunnel==0.4.0
starlette==0.47.2
stringzilla==3.12.6
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.6.0+cu124
torch-geometric==2.6.1
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.55.2
trio==0.30.0
trio-websocket==0.12.2
triton==3.2.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.179
ultralytics-thop==2.0.15
uncertainties==3.2.3
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
